@@ -26,9 +26,9 @@
 
 
     "require": {
-        "php": "~7.1",
+        "php": "~7.2",
 
-        "twig/twig": "~2.0"
+        "twig/twig": "~3.0"
     },
 
     "require-dev": {
